Refactor routesHash
